cli/azd/extensions/microsoft.azd.extensions/internal/resources/languages/python/requirements.txt
grpcio==1.71.0  
grpcio-tools==1.71.0  
protobuf==5.29.0
grpc-interceptor==0.15.4  
azure.identity== 1.21.0
azure.mgmt.resource== 23.3.0
rich==14.0.0